Refactored `load` & `download_bionemo_data` into bionemo-core
`bionemo-testing` re-exports the same values defined in `__all__`: they
are imported from `bionemo-core` as the implementations have moved.
Note that the tests have also moved.

Now, all sub-packages can use `load` at runtime, not just during tests.

All previous imports of `bionemo.testing.load` have been changed to
`bionemo.core.load`.

Additionally moves over the YAML resource files from bionemo-testing
into bionemo-core & adjusted the `get_all_resources` function.

This PR also fixes an error in the naming convention for
`bionemo-core`'s tests.
