+++ sub-packages/bionemo-core/src/bionemo/core/data/resources/esm2.yaml
