+++ sub-packages/bionemo-core/src/bionemo/core/data/resources/scdl.yaml
