+++ sub-packages/bionemo-core/src/bionemo/core/data/resources/single_cell.yaml
